move to btcstaking-helper
